+++ immunity-monitoring/files/lib/openwisp-monitoring/dhcp.lua
@@ -1,5 +1,5 @@
 -- retrieve dhcp leases
-local utils = require('openwisp-monitoring.utils')
+local utils = require('immunity-monitoring.utils')
 local uci = require('uci')
 local uci_cursor = uci.cursor()
 local io = require('io')