+++ immunity-monitoring/tests/test_neighbors.lua
@@ -1,5 +1,5 @@
 package.path = package.path ..
-                 ";../files/lib/openwisp-monitoring/?.lua;../files/sbin/?.lua"
+                 ";../files/lib/immunity-monitoring/?.lua;../files/sbin/?.lua"
 
 local luaunit = require('luaunit')
 local cjson = require('cjson')